2-Back_End/Semana-02/dia-04/requirements.txt
# pyrebase4: para integrar python con firebase
pyrebase4

# Flask: para implementar web con python utilizamos el microframerwork
Flask

# Se instala flask-bootstrap
flask-bootstrap

# Flask-wtf: Permite trabajar con formularios en flask
Flask-WTF

# python-decouple: sirve para poder utilizar variables de entorno
# python-decouple

# pythone-dotenv: sirve para poder utilizar variables de entorno
python-dotenv